# Conflicts:
#	README.md

+++ src/app/service/report-service/report.service.ts
@@ -1,7 +1,11 @@
 import {Injectable} from '@angular/core';
 import {PigReport} from "../../ts/PigReport";
 import {HttpClient} from "@angular/common/http";
-import {Observable} from "rxjs";
+import {Subject} from "rxjs";
+import "blueimp-md5";
+import {Browser} from "leaflet";
+import win = Browser.win;
+import md5 from "blueimp-md5";
 
 
 export interface Data{
@@ -20,8 +24,11 @@ export class ReportService {
   constructor(private http: HttpClient) { }
   reports : PigReport[] = [];
   url = `https://272.selfip.net/apps/yw1CIgsa3G/collections/pigs/documents/`;
-  newPigReport = new Observable<PigReport>;
-  deletedPigReport = new Observable<PigReport>;
+  newPigReportSource = new Subject<PigReport>();
+  newPigReportChanges$ = this.newPigReportSource.asObservable()
+  deletedPigReportSource = new Subject<PigReport>();
+  deletedPigReportChanges$ = this.deletedPigReportSource.asObservable()
+
 
 
 
@@ -33,27 +40,34 @@ export class ReportService {
     }
 
 
-    // return [
-    //   new PigReport(new Person("Sarah",0),new Pig(123,"Hock"),new PigLocation(-122.915667,49.278931,"SFU Burnaby"),Status.readyPickup, ""),
-    //   new PigReport(new Person("Darren",0),new Pig(123,"Hock"),new PigLocation(-122.915667,49.278931,"SFU Burnaby"),Status.readyPickup, ""),
-    //   new PigReport(new Person("Darrick",0),new Pig(123,"Hock"),new PigLocation(-122.915667,49.278931,"SFU Burnaby"),Status.readyPickup, ""),
-    //   new PigReport(new Person("John",0),new Pig(123,"Hock"),new PigLocation(-122.915667,49.278931,"SFU Burnaby"),Status.readyPickup, ""),
-    //   new PigReport(new Person("Timothy",0),new Pig(145,"Hock"),new PigLocation(-122.915667,49.278931,"SFU Burnaby"),Status.readyPickup, ""),
-    // ]
-
   deleteReport(report:PigReport): void{
-    console.log(`Report was sent for deletion`)
-    this.http.delete(this.url + report.key).subscribe((data:any)=>{
-      console.log(data);
-    })
+    console.log(`Report was sent for deletion`);
+    let windowPrompt = window.prompt("Type in password");
+    {
+        if (md5(windowPrompt!) == `84892b91ef3bf9d216bbc6e88d74a77c`) {
+          this.http.delete(this.url + report.key).subscribe((data: any) => {
+            console.log(data);
+          });
+          this.deletedPigReportSource.next(report);
+        }else{
+        window.alert(`Wrong password. Please try again`)}
+      }
   }
 
   addNewReport(report:PigReport) : void{
     console.log(`Report added:`);
-    console.log(report);
-    this.http.post<Data>(this.url,
-      {'key':report.key, 'data':report}).subscribe((data:any)=>{
-        console.log(data);
-    })
-  }
+    let windowPrompt = window.prompt("Type in password");
+    console.log(`api.hashify.net/hash/md5/hex?value=${windowPrompt}`);
+    {
+        if (md5(windowPrompt!) == `84892b91ef3bf9d216bbc6e88d74a77c`) {
+          this.newPigReportSource.next(report);
+          console.log(report);
+          this.http.post<Data>(this.url,
+            {'key':report.key, 'data':report}).subscribe((data:any)=>{
+              console.log(data);})
+        }
+        else{
+          window.alert(`Wrong password. Please try again`)}
+      }
+    }
 }
